+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Make `settings` page in admin panel lazy
 * Refactoring session service in admin-panel
 * Refactoring `server-info` page component in admin-panel
+* Reset password in admin-panel [#86](https://github.com/sham-ui/sham-ui-cli/issues/86)
 
 ## 4.3.0
 * Add template for `styled-component` [#3](https://github.com/sham-ui/sham-ui-cli/issues/3)